common.utils: automatic refactor: HTTP
Move `{http,api}_request{,_coro}` to `common.http`.
