+++ lrrbot/commands/live.py
@@ -1,3 +1,4 @@
+import common.http
 import lrrbot.decorators
 from lrrbot.main import bot
 from lrrbot import twitch
@@ -13,7 +14,7 @@ import irc.client
 @utils.cache(24 * 60 * 60)
 @asyncio.coroutine
 def extract_new_channels(loop):
-	data = yield from utils.http_request_coro(googlecalendar.EVENTS_URL % urllib.parse.quote(googlecalendar.CALENDAR_FAN), {
+	data = yield from common.http.http_request_coro(googlecalendar.EVENTS_URL % urllib.parse.quote(googlecalendar.CALENDAR_FAN), {
 		"key": config["google_key"],
 		"maxResults": 25000,
 	})
