+++ lrrbot/googlecalendar.py
@@ -6,6 +6,7 @@ import dateutil.parser
 import pytz
 import pytz.exceptions
 
+import common.http
 from common import utils
 from common.config import config
 
@@ -47,7 +48,7 @@ def get_upcoming_events(calendar, after=None):
 		"timeZone": config['timezone'].zone,
 		"key": config['google_key'],
 	}
-	res = utils.http_request(url, data)
+	res = common.http.http_request(url, data)
 	res = json.loads(res)
 	if 'error' in res:
 		raise Exception(res['error']['message'])
